pyproject.toml
[tool.poetry]
name = "aetheriadiscordbot"
version = "0.1.0"
description = ""
authors = ["AgentLoneStar <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord-pretty-help = "^1.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
